lint

+++ client/capi/example/examples/group.h
@@ -20,8 +20,8 @@
 void group(void* autd) {
   AUTDSetSilentMode(autd, true);
 
-  double x = TRANS_SPACING_MM * ((NUM_TRANS_X - 1) / 2.0);
-  double y = TRANS_SPACING_MM * ((NUM_TRANS_Y - 1) / 2.0);
+  double x = TRANS_SPACING_MM * (((double)NUM_TRANS_X - 1.0) / 2.0);
+  double y = TRANS_SPACING_MM * (((double)NUM_TRANS_Y - 1.0) / 2.0);
 
   void* g1 = NULL;
   AUTDGainFocalPoint(&g1, x, y, 150.0, 0xFF);
