+++ client/capi/example/examples/seq_gain.h
@@ -24,8 +24,8 @@
 void seq_gain(void* autd) {
   AUTDSetSilentMode(autd, false);
 
-  double x = TRANS_SPACING_MM * ((NUM_TRANS_X - 1) / 2.0);
-  double y = TRANS_SPACING_MM * ((NUM_TRANS_Y - 1) / 2.0);
+  double x = TRANS_SPACING_MM * (((double)NUM_TRANS_X - 1.0) / 2.0);
+  double y = TRANS_SPACING_MM * (((double)NUM_TRANS_Y - 1.0) / 2.0);
   double z = 150.0;
 
   void* seq = NULL;
